+++ src/components/Preview.tsx
@@ -8,11 +8,11 @@ import { useInView } from 'react-intersection-observer'
 
 import { Component } from '@/interface/component'
 
-import { transformComponentHtml } from '@/utils/componentHelpers'
-import { componentBreakpoints } from '@/utils/componentBreakpoints'
+import { componentHtml as transformComponentHtml } from '@/services/utils/transformers'
+import { componentBreakpoints } from '@/services/utils/breakpoints'
 
-import { useAppSelector } from '@/hooks/useStore'
-import { settingsState } from '@/store/slices/settings'
+import { useAppSelector } from '@/services/hooks/useStore'
+import { settingsState } from '@/services/store/slices/settings'
 
 import Breakpoint from '@/components/PreviewBreakpoint'
 import Code from '@/components/PreviewCode'
