+++ src/components/SettingComponentLinks.tsx
@@ -1,5 +1,5 @@
-import { useAppSelector, useAppDispatch } from '@/hooks/useStore'
-import { toggleLinks, settingsState } from '@/store/slices/settings'
+import { useAppSelector, useAppDispatch } from '@/services/hooks/useStore'
+import { toggleLinks, settingsState } from '@/services/store/slices/settings'
 
 import SettingTitle from '@/components/SettingTitle'
 import SettingToggle from '@/components/SettingToggle'
