+++ src/components/SettingTitle.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react'
 import { useRouter } from 'next/router'
 
 type Props = {
