--- src/utils/menuLinks.ts
@@ -1,29 +1,0 @@
-import { MenuLink } from '../interface/global'
-
-export const menuLinks: Array<MenuLink> = [
-  {
-    title: 'Marketing',
-    href: '/components/marketing',
-    external: false,
-  },
-  {
-    title: 'eCommerce',
-    href: '/components/ecommerce',
-    external: false,
-  },
-  {
-    title: 'Application UI',
-    href: '/components/application-ui',
-    external: false,
-  },
-  {
-    title: 'Blog',
-    href: '/blog',
-    external: false,
-  },
-  {
-    title: 'HyperJS',
-    href: 'https://js.hyperui.dev',
-    external: true,
-  },
-]
